requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
jinja2==3.1.3
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
aiosqlite==0.19.0

# HTTP Clients
httpx==0.26.0
aiohttp==3.9.1

# Data Validation
pydantic==2.5.3

# PDF Reporting
reportlab==4.0.9

# Utilities
python-dotenv==1.0.0
